YW2-227 feat: 어댑터 반영
